Add initial implementation of a ResourceParser tool
The tool was redone in C++ to make development a bit easier. It takes
a binary resource file (BIN1) and its type (eg. TEMP) as an input,
and it produces a JSON representation of that file.

This depends on code generated by Statman (which is committed into
the `Src/Generated` folder) but there's still a few things missing
and Statman is not yet able to generate code for all types.