[ResourceLib] Add support for generating "compatible" resources
When using the "Compatible" variants of the ResourceGenerator methods,
ResourceLib will attempt to generate resources that more closely resemble
the ones originally found in the game (identical in many cases). The
differences between generating "compatible" resources or not doesn't
really matter when one wants to use them in a game, but some third
party tools might play better with the generated resources when using
this mode.

The main differences are as follows:

1. ResourceLib will try to deduplicate ZVariants in-memory, so multiple
ZVariants that have equal data actually point to the same data in-memory.
This can make files smaller and closer to the original ones, but it also
makes generation take significantly more time.

2. ResourceLib will prepend a 32-bit integer to every serialized string
with the length of the string (including its null terminator).

3. ResourceLib will prepend a 32-bit integer to every serialized array
with the number of elements in the array.