Fix encrypted RTLV serialization
Was previously using ZString, which adds null terminators to all strings
during serialization because the engine sometimes explodes when they
are not. In this case however, we want the opposite, and our encrypted
strings should not have a null terminator.

Another weird issue is that even though we have an array of
ZRuntimeResourceIDs, if we generate a BIN1 segment with the offsets
of those rrids like we normally would for any other data type, the game
will crash as soon as we try to play the relevant video. In order to
avoid this issue, we serialize the ZRuntimeResourceIDs as just plain
data, skipping the creation of this segment.

Fixes #23